Refactor cockpit-ui setup and add support for SSL
This renames the role 'cockpit-ui' to
'openshift_hosted_registry_console' to bring it in line with
the other roles related to openshift_hosted.

This is a refactor to significantly improve the configurability
of the registry console service and route in order to bring it
in line with the way we configure docker-registry.

Effectively, this adds the ability to select a hostname for the
registry console and adds support for securing the registry-console
route with a provided SSL certificate either with passthrough or
reencrypt termination.

We maintain backwards compatibility by keeping the same default
which provides a default registry-console hostname and self-signed
certificates on a passthrough route.

(cherry picked from commit 6cda803b367b8edc05120f3e9e018cd7167604cc)
(cherry picked from commit 400bad5464b9cfcb05166eaa8cb09f66343380b2)